+++ src/main/java/net/silentchaos512/gear/parts/LazyPartData.java
@@ -2,9 +2,12 @@ package net.silentchaos512.gear.parts;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
+import net.minecraft.item.ItemStack;
+import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.util.JSONUtils;
 import net.minecraft.util.ResourceLocation;
 import net.silentchaos512.gear.api.parts.IGearPart;
+import net.silentchaos512.gear.api.parts.IPartData;
 import net.silentchaos512.gear.api.parts.MaterialGrade;
 
 import javax.annotation.Nullable;
@@ -16,7 +19,7 @@ import java.util.stream.Collectors;
  * A "lazy" version of {@link PartData}. Since {@link IGearPart}s may not exist when certain things
  * like loot tables are loaded, {@code LazyPartData} can be used to represent a future part.
  */
-public class LazyPartData {
+public class LazyPartData implements IPartData {
     private final ResourceLocation partId;
     private final MaterialGrade grade;
 
@@ -29,13 +32,37 @@ public class LazyPartData {
         this.grade = grade;
     }
 
+    @Override
+    public ResourceLocation getPartId() {
+        return partId;
+    }
+
     @Nullable
+    @Override
     public IGearPart getPart() {
         return PartManager.get(partId);
     }
 
+    @Override
     public MaterialGrade getGrade() {
         return grade;
+    }
+
+    @Override
+    public ItemStack getCraftingItem() {
+        IGearPart part = getPart();
+        if (part != null)
+            return PartData.of(part).getCraftingItem();
+        return ItemStack.EMPTY;
+    }
+
+    @Override
+    public CompoundNBT write(CompoundNBT tags) {
+        tags.putString("ID", partId.toString());
+        if (this.grade != MaterialGrade.NONE) {
+            tags.putString("Grade", this.grade.name());
+        }
+        return tags;
     }
 
     public boolean isValid() {