+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Mine and Slash random drops are now tier 1 minimum
 - Mine and Slash random drops should now be salvagable
+### Fixed
+- JEI failing to load on servers [#52]
 
 ## [1.3.8] - 2019-08-13
 ### Fixed